@@ -121,57 +135,70 @@ dependencies:
   path_provider_linux: 2.2.1
   path_provider_platform_interface: 2.1.2
   path_provider_windows: 2.3.0
+  pdfx: 2.9.2
   plugin_platform_interface: 2.1.8
+  postgrest: 2.4.2
   progressive_time_picker: ^1.0.3
-  provider: 6.1.2
+  provider: 6.1.5
+  realtime_client: 2.5.0
   rxdart: 0.27.7
   screenshot: 3.0.0
-  shared_preferences: 2.3.2
-  shared_preferences_android: 2.3.2
-  shared_preferences_foundation: 2.5.2
+  shared_preferences: 2.5.3
+  shared_preferences_android: 2.4.10
+  shared_preferences_foundation: 2.5.4
   shared_preferences_linux: 2.4.1
   shared_preferences_platform_interface: 2.4.1
-  shared_preferences_web: 2.4.2
+  shared_preferences_web: 2.4.3
   shared_preferences_windows: 2.4.1
-  sign_in_with_apple: 6.1.2
-  sign_in_with_apple_platform_interface: 1.1.0
-  sign_in_with_apple_web: 2.1.0
+  sign_in_with_apple: 7.0.1
+  sign_in_with_apple_platform_interface: 2.0.0
+  sign_in_with_apple_web: 3.0.0
   signature: 5.4.0
   smooth_page_indicator: 1.1.0
   sqflite: 2.3.3+1
   sqflite_common: 2.5.4+3
   stop_watch_timer: 3.0.2
+  storage_client: 2.4.0
   stream_transform: 2.1.0
-  syncfusion_flutter_charts: ^20.4.54
-  syncfusion_flutter_gauges: ^20.1.59
-  synchronized: 3.2.0
+  supabase: 2.7.0
+  supabase_flutter: 2.9.0
+  syncfusion_flutter_charts: ^30.1.39
+  syncfusion_flutter_gauges: ^30.1.39
+  syncfusion_flutter_pdfviewer: ^30.1.39
+  synchronized: 3.4.0
   time_picker_spinner_pop_up: ^2.0.0
   time_planner: ^0.1.2+1
-  timeago: 3.6.1
-  url_launcher: 6.3.0
-  url_launcher_android: 6.3.10
-  url_launcher_ios: 6.3.1
-  url_launcher_linux: 3.2.0
-  url_launcher_macos: 3.2.1
+  timeago: 3.7.1
+  url_launcher: 6.3.1
+  url_launcher_android: 6.3.16
+  url_launcher_ios: 6.3.3
+  url_launcher_linux: 3.2.1
+  url_launcher_macos: 3.2.2
   url_launcher_platform_interface: 2.3.2
-  url_launcher_web: 2.3.3
-  url_launcher_windows: 3.1.2
-  video_player: 2.9.2
-  video_player_android: 2.7.13
-  video_player_avfoundation: 2.6.2
-  video_player_platform_interface: 6.2.3
-  video_player_web: 2.3.2
-  webview_flutter: ^4.8.0
+  url_launcher_web: 2.4.1
+  url_launcher_windows: 3.1.4
+  video_player: 2.10.0
+  video_player_android: 2.8.7
+  video_player_avfoundation: 2.7.1
+  video_player_platform_interface: 6.3.0
+  video_player_web: 2.3.5
+  wakelock_plus: ^1.3.2
+  webview_flutter: 4.13.0
+  webview_flutter_android: 4.7.0
+  webview_flutter_platform_interface: 2.13.1
+  webview_flutter_wkwebview: 3.22.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.0
 
 dependency_overrides:
-  http: 1.2.2
-  intl: 0.19.0
+  device_info_plus: 11.5.0
+  device_info_plus_platform_interface: 7.0.3
+  http: 1.4.0
+  intl: 0.20.2
+  rxdart: 0.27.7
   uuid: ^4.0.0
-  win32: 5.5.1
 
 dev_dependencies:
   flutter_launcher_icons: 0.13.1
