@@ -11,6 +11,7 @@ import '/index.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:google_fonts/google_fonts.dart';
+import 'package:provider/provider.dart';
 import 'admin_seat_configure_model.dart';
 export 'admin_seat_configure_model.dart';
 
@@ -56,7 +57,7 @@ class _AdminSeatConfigureWidgetState extends State<AdminSeatConfigureWidget> {
           },
         );
 
-        context.goNamed(HomePageWidget.routeName);
+        context.goNamed(HomePageCopyWidget.routeName);
       }
     });
 
@@ -72,6 +73,8 @@ class _AdminSeatConfigureWidgetState extends State<AdminSeatConfigureWidget> {
 
   @override
   Widget build(BuildContext context) {
+    context.watch<FFAppState>();
+
     return GestureDetector(
       onTap: () {
         FocusScope.of(context).unfocus();
@@ -285,55 +288,85 @@ class _AdminSeatConfigureWidgetState extends State<AdminSeatConfigureWidget> {
                                                   padding: EdgeInsetsDirectional
                                                       .fromSTEB(
                                                           8.0, 4.0, 8.0, 4.0),
-                                                  child: FutureBuilder<int>(
-                                                    future:
-                                                        queryPlannerVariableListRecordCount(
-                                                      queryBuilder:
-                                                          (plannerVariableListRecord) =>
-                                                              plannerVariableListRecord
-                                                                  .where(
-                                                                    'plannerSubmitted',
-                                                                    isEqualTo:
-                                                                        true,
-                                                                  )
-                                                                  .where(
-                                                                    'adminChecked',
-                                                                    isNotEqualTo:
-                                                                        true,
-                                                                  ),
-                                                    ),
-                                                    builder:
-                                                        (context, snapshot) {
-                                                      // Customize what your widget looks like when it's loading.
-                                                      if (!snapshot.hasData) {
-                                                        return Center(
-                                                          child: SizedBox(
-                                                            width: 50.0,
-                                                            height: 50.0,
-                                                            child:
-                                                                CircularProgressIndicator(
-                                                              valueColor:
-                                                                  AlwaysStoppedAnimation<
-                                                                      Color>(
-                                                                FlutterFlowTheme.of(
-                                                                        context)
-                                                                    .primary,
+                                                  child: AuthUserStreamWidget(
+                                                    builder: (context) =>
+                                                        FutureBuilder<int>(
+                                                      future:
+                                                          queryPlannerVariableListRecordCount(
+                                                        queryBuilder:
+                                                            (plannerVariableListRecord) =>
+                                                                plannerVariableListRecord
+                                                                    .where(
+                                                                      'plannerSubmitted',
+                                                                      isEqualTo:
+                                                                          true,
+                                                                    )
+                                                                    .where(
+                                                                      'adminChecked',
+                                                                      isNotEqualTo:
+                                                                          true,
+                                                                    )
+                                                                    .where(
+                                                                      'submittedDate',
+                                                                      isEqualTo:
+                                                                          FFAppState()
+                                                                              .plannerDateSelected,
+                                                                    )
+                                                                    .where(
+                                                                      'basicInfo.userSpot',
+                                                                      isEqualTo: valueOrDefault(currentUserDocument?.spot, '') !=
+                                                                              ''
+                                                                          ? valueOrDefault(
+                                                                              currentUserDocument?.spot,
+                                                                              '')
+                                                                          : null,
+                                                                    ),
+                                                      ),
+                                                      builder:
+                                                          (context, snapshot) {
+                                                        // Customize what your widget looks like when it's loading.
+                                                        if (!snapshot.hasData) {
+                                                          return Center(
+                                                            child: SizedBox(
+                                                              width: 50.0,
+                                                              height: 50.0,
+                                                              child:
+                                                                  CircularProgressIndicator(
+                                                                valueColor:
+                                                                    AlwaysStoppedAnimation<
+                                                                        Color>(
+                                                                  FlutterFlowTheme.of(
+                                                                          context)
+                                                                      .primary,
+                                                                ),
                                                               ),
                                                             ),
-                                                          ),
-                                                        );
-                                                      }
-                                                      int textCount =
-                                                          snapshot.data!;
+                                                          );
+                                                        }
+                                                        int textCount =
+                                                            snapshot.data!;
 
-                                                      return Text(
-                                                        textCount.toString(),
-                                                        style: FlutterFlowTheme
-                                                                .of(context)
-                                                            .bodyMedium
-                                                            .override(
-                                                              font: GoogleFonts
-                                                                  .plusJakartaSans(
+                                                        return Text(
+                                                          textCount.toString(),
+                                                          style: FlutterFlowTheme
+                                                                  .of(context)
+                                                              .bodyMedium
+                                                              .override(
+                                                                font: GoogleFonts
+                                                                    .plusJakartaSans(
+                                                                  fontWeight:
+                                                                      FontWeight
+                                                                          .w500,
+                                                                  fontStyle: FlutterFlowTheme.of(
+                                                                          context)
+                                                                      .bodyMedium
+                                                                      .fontStyle,
+                                                                ),
+                                                                color: Colors
+                                                                    .white,
+                                                                fontSize: 14.0,
+                                                                letterSpacing:
+                                                                    0.0,
                                                                 fontWeight:
                                                                     FontWeight
                                                                         .w500,
@@ -342,22 +375,9 @@ class _AdminSeatConfigureWidgetState extends State<AdminSeatConfigureWidget> {
                                                                     .bodyMedium
                                                                     .fontStyle,
                                                               ),
-                                                              color:
-                                                                  Colors.white,
-                                                              fontSize: 14.0,
-                                                              letterSpacing:
-                                                                  0.0,
-                                                              fontWeight:
-                                                                  FontWeight
-                                                                      .w500,
-                                                              fontStyle:
-                                                                  FlutterFlowTheme.of(
-                                                                          context)
-                                                                      .bodyMedium
-                                                                      .fontStyle,
-                                                            ),
-                                                      );
-                                                    },
+                                                        );
+                                                      },
+                                                    ),
                                                   ),
                                                 ),
                                               ),
@@ -396,49 +416,194 @@ class _AdminSeatConfigureWidgetState extends State<AdminSeatConfigureWidget> {
                                             8.0, 0.0, 6.0, 0.0),
                                         child: Row(
                                           mainAxisSize: MainAxisSize.max,
+                                          mainAxisAlignment:
+                                              MainAxisAlignment.spaceBetween,
                                           children: [
-                                            Icon(
-                                              Icons.space_dashboard_rounded,
-                                              color: Color(0xFF15161E),
-                                              size: 24.0,
-                                            ),
-                                            Padding(
-                                              padding: EdgeInsetsDirectional
-                                                  .fromSTEB(
-                                                      12.0, 0.0, 0.0, 0.0),
-                                              child: Text(
-                                                FFLocalizations.of(context)
-                                                    .getText(
-                                                  'zxv0talx' /* 시간표 관리 */,
-                                                ),
-                                                style:
-                                                    FlutterFlowTheme.of(context)
-                                                        .bodyMedium
-                                                        .override(
-                                                          font: GoogleFonts
-                                                              .plusJakartaSans(
-                                                            fontWeight:
-                                                                FontWeight.w500,
-                                                            fontStyle:
-                                                                FlutterFlowTheme.of(
+                                            Row(
+                                              mainAxisSize: MainAxisSize.max,
+                                              children: [
+                                                Row(
+                                                  mainAxisSize:
+                                                      MainAxisSize.max,
+                                                  children: [
+                                                    Icon(
+                                                      Icons
+                                                          .space_dashboard_rounded,
+                                                      color: Color(0xFF15161E),
+                                                      size: 24.0,
+                                                    ),
+                                                    Padding(
+                                                      padding:
+                                                          EdgeInsetsDirectional
+                                                              .fromSTEB(
+                                                                  12.0,
+                                                                  0.0,
+                                                                  0.0,
+                                                                  0.0),
+                                                      child: Text(
+                                                        FFLocalizations.of(
+                                                                context)
+                                                            .getText(
+                                                          'obzevtao' /* 시간표 관리 */,
+                                                        ),
+                                                        style: FlutterFlowTheme
+                                                                .of(context)
+                                                            .bodyMedium
+                                                            .override(
+                                                              font: GoogleFonts
+                                                                  .plusJakartaSans(
+                                                                fontWeight:
+                                                                    FontWeight
+                                                                        .w500,
+                                                                fontStyle: FlutterFlowTheme.of(
                                                                         context)
                                                                     .bodyMedium
                                                                     .fontStyle,
-                                                          ),
-                                                          color:
-                                                              Color(0xFF15161E),
-                                                          fontSize: 14.0,
-                                                          letterSpacing: 0.0,
-                                                          fontWeight:
-                                                              FontWeight.w500,
-                                                          fontStyle:
+                                                              ),
+                                                              color: Color(
+                                                                  0xFF15161E),
+                                                              fontSize: 14.0,
+                                                              letterSpacing:
+                                                                  0.0,
+                                                              fontWeight:
+                                                                  FontWeight
+                                                                      .w500,
+                                                              fontStyle:
+                                                                  FlutterFlowTheme.of(
+                                                                          context)
+                                                                      .bodyMedium
+                                                                      .fontStyle,
+                                                            ),
+                                                      ),
+                                                    ),
+                                                  ],
+                                                ),
+                                              ],
+                                            ),
+                                            if (true /* Warning: Trying to access variable not yet defined. */)
+                                              AuthUserStreamWidget(
+                                                builder: (context) =>
+                                                    FutureBuilder<int>(
+                                                  future: queryUsersRecordCount(
+                                                    queryBuilder:
+                                                        (usersRecord) =>
+                                                            usersRecord
+                                                                .where(
+                                                                  'scheduleChangeRequest',
+                                                                  isEqualTo:
+                                                                      true,
+                                                                )
+                                                                .where(
+                                                                  'spot',
+                                                                  isEqualTo: valueOrDefault(
+                                                                              currentUserDocument
+                                                                                  ?.spot,
+                                                                              '') !=
+                                                                          ''
+                                                                      ? valueOrDefault(
+                                                                          currentUserDocument
+                                                                              ?.spot,
+                                                                          '')
+                                                                      : null,
+                                                                )
+                                                                .where(
+                                                                  'seatNo',
+                                                                  isGreaterThan:
+                                                                      0,
+                                                                )
+                                                                .where(
+                                                                  'seatNo',
+                                                                  isLessThan:
+                                                                      200,
+                                                                ),
+                                                  ),
+                                                  builder: (context, snapshot) {
+                                                    // Customize what your widget looks like when it's loading.
+                                                    if (!snapshot.hasData) {
+                                                      return Center(
+                                                        child: SizedBox(
+                                                          width: 50.0,
+                                                          height: 50.0,
+                                                          child:
+                                                              CircularProgressIndicator(
+                                                            valueColor:
+                                                                AlwaysStoppedAnimation<
+                                                                    Color>(
                                                               FlutterFlowTheme.of(
                                                                       context)
-                                                                  .bodyMedium
-                                                                  .fontStyle,
+                                                                  .primary,
+                                                            ),
+                                                          ),
                                                         ),
+                                                      );
+                                                    }
+                                                    int containerCount =
+                                                        snapshot.data!;
+
+                                                    return Container(
+                                                      width: 32.0,
+                                                      height: 32.0,
+                                                      decoration: BoxDecoration(
+                                                        color:
+                                                            Color(0xFFFF0000),
+                                                        borderRadius:
+                                                            BorderRadius
+                                                                .circular(8.0),
+                                                      ),
+                                                      child: Align(
+                                                        alignment:
+                                                            AlignmentDirectional(
+                                                                0.0, 0.0),
+                                                        child: Padding(
+                                                          padding:
+                                                              EdgeInsetsDirectional
+                                                                  .fromSTEB(
+                                                                      8.0,
+                                                                      4.0,
+                                                                      8.0,
+                                                                      4.0),
+                                                          child: Text(
+                                                            valueOrDefault<
+                                                                String>(
+                                                              containerCount
+                                                                  .toString(),
+                                                              '1',
+                                                            ),
+                                                            style: FlutterFlowTheme
+                                                                    .of(context)
+                                                                .bodyMedium
+                                                                .override(
+                                                                  font: GoogleFonts
+                                                                      .plusJakartaSans(
+                                                                    fontWeight:
+                                                                        FontWeight
+                                                                            .w500,
+                                                                    fontStyle: FlutterFlowTheme.of(
+                                                                            context)
+                                                                        .bodyMedium
+                                                                        .fontStyle,
+                                                                  ),
+                                                                  color: Colors
+                                                                      .white,
+                                                                  fontSize:
+                                                                      14.0,
+                                                                  letterSpacing:
+                                                                      0.0,
+                                                                  fontWeight:
+                                                                      FontWeight
+                                                                          .w500,
+                                                                  fontStyle: FlutterFlowTheme.of(
+                                                                          context)
+                                                                      .bodyMedium
+                                                                      .fontStyle,
+                                                                ),
+                                                          ),
+                                                        ),
+                                                      ),
+                                                    );
+                                                  },
+                                                ),
                                               ),
-                                            ),
                                           ],
                                         ),
                                       ),
@@ -707,46 +872,74 @@ class _AdminSeatConfigureWidgetState extends State<AdminSeatConfigureWidget> {
                                                 padding: EdgeInsetsDirectional
                                                     .fromSTEB(
                                                         8.0, 4.0, 8.0, 4.0),
-                                                child: FutureBuilder<int>(
-                                                  future: queryUsersRecordCount(
-                                                    queryBuilder:
-                                                        (usersRecord) =>
-                                                            usersRecord.where(
-                                                      'seatRequest',
-                                                      isEqualTo: true,
+                                                child: AuthUserStreamWidget(
+                                                  builder: (context) =>
+                                                      FutureBuilder<int>(
+                                                    future:
+                                                        queryUsersRecordCount(
+                                                      queryBuilder:
+                                                          (usersRecord) =>
+                                                              usersRecord
+                                                                  .where(
+                                                                    'seatRequest',
+                                                                    isEqualTo:
+                                                                        true,
+                                                                  )
+                                                                  .where(
+                                                                    'spot',
+                                                                    isEqualTo: valueOrDefault(currentUserDocument?.spot, '') !=
+                                                                            ''
+                                                                        ? valueOrDefault(
+                                                                            currentUserDocument?.spot,
+                                                                            '')
+                                                                        : null,
+                                                                  ),
                                                     ),
-                                                  ),
-                                                  builder: (context, snapshot) {
-                                                    // Customize what your widget looks like when it's loading.
-                                                    if (!snapshot.hasData) {
-                                                      return Center(
-                                                        child: SizedBox(
-                                                          width: 50.0,
-                                                          height: 50.0,
-                                                          child:
-                                                              CircularProgressIndicator(
-                                                            valueColor:
-                                                                AlwaysStoppedAnimation<
-                                                                    Color>(
-                                                              FlutterFlowTheme.of(
-                                                                      context)
-                                                                  .primary,
+                                                    builder:
+                                                        (context, snapshot) {
+                                                      // Customize what your widget looks like when it's loading.
+                                                      if (!snapshot.hasData) {
+                                                        return Center(
+                                                          child: SizedBox(
+                                                            width: 50.0,
+                                                            height: 50.0,
+                                                            child:
+                                                                CircularProgressIndicator(
+                                                              valueColor:
+                                                                  AlwaysStoppedAnimation<
+                                                                      Color>(
+                                                                FlutterFlowTheme.of(
+                                                                        context)
+                                                                    .primary,
+                                                              ),
                                                             ),
                                                           ),
-                                                        ),
-                                                      );
-                                                    }
-                                                    int textCount =
-                                                        snapshot.data!;
+                                                        );
+                                                      }
+                                                      int textCount =
+                                                          snapshot.data!;
 
-                                                    return Text(
-                                                      textCount.toString(),
-                                                      style: FlutterFlowTheme
-                                                              .of(context)
-                                                          .bodyMedium
-                                                          .override(
-                                                            font: GoogleFonts
-                                                                .plusJakartaSans(
+                                                      return Text(
+                                                        textCount.toString(),
+                                                        style: FlutterFlowTheme
+                                                                .of(context)
+                                                            .bodyMedium
+                                                            .override(
+                                                              font: GoogleFonts
+                                                                  .plusJakartaSans(
+                                                                fontWeight:
+                                                                    FontWeight
+                                                                        .w500,
+                                                                fontStyle: FlutterFlowTheme.of(
+                                                                        context)
+                                                                    .bodyMedium
+                                                                    .fontStyle,
+                                                              ),
+                                                              color:
+                                                                  Colors.white,
+                                                              fontSize: 14.0,
+                                                              letterSpacing:
+                                                                  0.0,
                                                               fontWeight:
                                                                   FontWeight
                                                                       .w500,
@@ -756,19 +949,9 @@ class _AdminSeatConfigureWidgetState extends State<AdminSeatConfigureWidget> {
                                                                       .bodyMedium
                                                                       .fontStyle,
                                                             ),
-                                                            color: Colors.white,
-                                                            fontSize: 14.0,
-                                                            letterSpacing: 0.0,
-                                                            fontWeight:
-                                                                FontWeight.w500,
-                                                            fontStyle:
-                                                                FlutterFlowTheme.of(
-                                                                        context)
-                                                                    .bodyMedium
-                                                                    .fontStyle,
-                                                          ),
-                                                    );
-                                                  },
+                                                      );
+                                                    },
+                                                  ),
                                                 ),
                                               ),
                                             ),
@@ -897,7 +1080,7 @@ class _AdminSeatConfigureWidgetState extends State<AdminSeatConfigureWidget> {
                                               child: Text(
                                                 FFLocalizations.of(context)
                                                     .getText(
-                                                  '3i4ft0rt' /* 포인트 관리 */,
+                                                  '3i4ft0rt' /* 포인트 및 벌점 관리 */,
                                                 ),
                                                 style:
                                                     FlutterFlowTheme.of(context)
@@ -1058,6 +1241,84 @@ class _AdminSeatConfigureWidgetState extends State<AdminSeatConfigureWidget> {
                                                 FFLocalizations.of(context)
                                                     .getText(
                                                   '3b5a1690' /* 입학원서/서약서확인 */,
+                                                ),
+                                                style:
+                                                    FlutterFlowTheme.of(context)
+                                                        .bodyMedium
+                                                        .override(
+                                                          font: GoogleFonts
+                                                              .plusJakartaSans(
+                                                            fontWeight:
+                                                                FontWeight.w500,
+                                                            fontStyle:
+                                                                FlutterFlowTheme.of(
+                                                                        context)
+                                                                    .bodyMedium
+                                                                    .fontStyle,
+                                                          ),
+                                                          color:
+                                                              Color(0xFF15161E),
+                                                          fontSize: 14.0,
+                                                          letterSpacing: 0.0,
+                                                          fontWeight:
+                                                              FontWeight.w500,
+                                                          fontStyle:
+                                                              FlutterFlowTheme.of(
+                                                                      context)
+                                                                  .bodyMedium
+                                                                  .fontStyle,
+                                                        ),
+                                              ),
+                                            ),
+                                          ],
+                                        ),
+                                      ),
+                                    ),
+                                  ),
+                                ),
+                                Padding(
+                                  padding: EdgeInsetsDirectional.fromSTEB(
+                                      16.0, 0.0, 16.0, 0.0),
+                                  child: InkWell(
+                                    splashColor: Colors.transparent,
+                                    focusColor: Colors.transparent,
+                                    hoverColor: Colors.transparent,
+                                    highlightColor: Colors.transparent,
+                                    onTap: () async {
+                                      context.pushNamed(
+                                          AdminPwCheckWidget.routeName);
+                                    },
+                                    child: AnimatedContainer(
+                                      duration: Duration(milliseconds: 200),
+                                      curve: Curves.easeInOut,
+                                      width: double.infinity,
+                                      height: 44.0,
+                                      decoration: BoxDecoration(
+                                        borderRadius:
+                                            BorderRadius.circular(12.0),
+                                        shape: BoxShape.rectangle,
+                                      ),
+                                      child: Padding(
+                                        padding: EdgeInsetsDirectional.fromSTEB(
+                                            8.0, 0.0, 6.0, 0.0),
+                                        child: Row(
+                                          mainAxisSize: MainAxisSize.max,
+                                          children: [
+                                            Icon(
+                                              Icons.work,
+                                              color:
+                                                  FlutterFlowTheme.of(context)
+                                                      .primaryText,
+                                              size: 24.0,
+                                            ),
+                                            Padding(
+                                              padding: EdgeInsetsDirectional
+                                                  .fromSTEB(
+                                                      12.0, 0.0, 0.0, 0.0),
+                                              child: Text(
+                                                FFLocalizations.of(context)
+                                                    .getText(
+                                                  'o7kxndhl' /* 아이디 / 비밀번호 관리 */,
                                                 ),
                                                 style:
                                                     FlutterFlowTheme.of(context)
