@@ -58,8 +58,30 @@ class GetActiveUsersAndDetailsByDateRangeCloudFunctionCallResponse {
   dynamic data;
 }
 
-class GenerateFeedbackOnCreateCloudFunctionCallResponse {
-  GenerateFeedbackOnCreateCloudFunctionCallResponse({
+class UpdateUserDataCloudFunctionCallResponse {
+  UpdateUserDataCloudFunctionCallResponse({
+    this.errorCode,
+    this.succeeded,
+    this.jsonBody,
+  });
+  String? errorCode;
+  bool? succeeded;
+  dynamic jsonBody;
+}
+
+class NewCloudFunctionCloudFunctionCallResponse {
+  NewCloudFunctionCloudFunctionCallResponse({
+    this.errorCode,
+    this.succeeded,
+    this.jsonBody,
+  });
+  String? errorCode;
+  bool? succeeded;
+  dynamic jsonBody;
+}
+
+class UpdatePointsOnAdminCheck4CloudFunctionCallResponse {
+  UpdatePointsOnAdminCheck4CloudFunctionCallResponse({
     this.errorCode,
     this.succeeded,
     this.jsonBody,
