@@ -1,7 +1,10 @@
+import '/auth/firebase_auth/auth_util.dart';
+import '/backend/backend.dart';
 import '/flutter_flow/flutter_flow_icon_button.dart';
 import '/flutter_flow/flutter_flow_theme.dart';
 import '/flutter_flow/flutter_flow_util.dart';
 import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
 import 'subject_bottom_sheet_model.dart';
 export 'subject_bottom_sheet_model.dart';
 
@@ -49,6 +52,8 @@ class _SubjectBottomSheetWidgetState extends State<SubjectBottomSheetWidget> {
 
   @override
   Widget build(BuildContext context) {
+    context.watch<FFAppState>();
+
     return Container(
       width: 350.0,
       height: 100.0,
@@ -175,7 +180,22 @@ class _SubjectBottomSheetWidgetState extends State<SubjectBottomSheetWidget> {
                   (e) => e..subject = _model.textController.text,
                 );
                 FFAppState().update(() {});
-                await Future.delayed(const Duration(milliseconds: 300));
+                await Future.delayed(
+                  Duration(
+                    milliseconds: 300,
+                  ),
+                );
+
+                await currentUserReference!.update({
+                  ...mapToFirestore(
+                    {
+                      'personalSubjectInfo':
+                          getPersonalSubjectEachListFirestoreData(
+                        FFAppState().personalSubjectInfo,
+                      ),
+                    },
+                  ),
+                });
                 Navigator.pop(context);
               },
             ),
