@@ -2,6 +2,7 @@
 import '/backend/backend.dart';
 import '/backend/schema/structs/index.dart';
 import '/backend/schema/enums/enums.dart';
+import '/backend/supabase/supabase.dart';
 import '/actions/actions.dart' as action_blocks;
 import '/flutter_flow/flutter_flow_theme.dart';
 import '/flutter_flow/flutter_flow_util.dart';
@@ -12,7 +13,7 @@ import 'package:flutter/material.dart';
 // Begin custom widget code
 // DO NOT REMOVE OR MODIFY THE CODE ABOVE!
 
-import 'package:webview_flutter/webview_flutter.dart';
+import 'package:flutter_inappwebview/flutter_inappwebview.dart';
 
 class FilteredWebView extends StatefulWidget {
   const FilteredWebView({
@@ -31,47 +32,46 @@ class FilteredWebView extends StatefulWidget {
 }
 
 class _FilteredWebViewState extends State<FilteredWebView> {
-  late final WebViewController _controller;
-
-  @override
-  void initState() {
-    super.initState();
-
-    _controller = WebViewController()
-      ..setJavaScriptMode(JavaScriptMode.unrestricted)
-      ..setNavigationDelegate(
-        NavigationDelegate(
-          onNavigationRequest: (request) {
-            final uri = Uri.parse(request.url);
-
-            // 허용할 도메인 및 경로
-            final isAllowedHost = uri.host == 'pf.kakao.com';
-            final isAllowedPath = uri.path.startsWith('/_cLWxgxj');
-
-            if (isAllowedHost && isAllowedPath) {
-              return NavigationDecision.navigate;
-            } else {
-              ScaffoldMessenger.of(context).showSnackBar(
-                const SnackBar(content: Text('허용되지 않은 링크입니다.')),
-              );
-              return NavigationDecision.prevent;
-            }
-          },
-        ),
-      )
-      ..loadRequest(
-          Uri.parse(widget.initialUrl ?? 'https://pf.kakao.com/_cLWxgxj'));
-  }
+  InAppWebViewController? _webViewController;
 
   @override
   Widget build(BuildContext context) {
     return SizedBox(
       width: widget.width,
       height: widget.height,
-      child: WebViewWidget(controller: _controller),
+      child: InAppWebView(
+        initialUrlRequest: URLRequest(
+          url: WebUri(
+            widget.initialUrl ?? 'https://pf.kakao.com/_cLWxgxj',
+          ),
+        ),
+        initialSettings: InAppWebViewSettings(
+          javaScriptEnabled: true,
+          // ✅ 버전 6.1.x에서는 MixedContentMode (공통 enum) 사용!
+          mixedContentMode: MixedContentMode.MIXED_CONTENT_ALWAYS_ALLOW,
+        ),
+        onWebViewCreated: (controller) {
+          _webViewController = controller;
+        },
+        shouldOverrideUrlLoading: (controller, navigationAction) async {
+          final uri = navigationAction.request.url;
+
+          final isAllowedHost =
+              uri?.host == 'pf.kakao.com' || uri?.host == 'k.kakaocdn.net';
+          final isAllowedPath = uri?.host == 'pf.kakao.com'
+              ? (uri?.path.startsWith('/_cLWxgxj') ?? false)
+              : true;
+
+          if (isAllowedHost && isAllowedPath) {
+            return NavigationActionPolicy.ALLOW;
+          } else {
+            ScaffoldMessenger.of(context).showSnackBar(
+              const SnackBar(content: Text('허용되지 않은 링크입니다.')),
+            );
+            return NavigationActionPolicy.CANCEL;
+          }
+        },
+      ),
     );
   }
 }
-
-// Set your widget name, define your parameter, and then add the
-// boilerplate code using the green button on the right!
