@@ -21,12 +21,14 @@ class SelectedFile {
     required this.bytes,
     this.dimensions,
     this.blurHash,
+    this.originalFilename = '',
   });
   final String storagePath;
   final String? filePath;
   final Uint8List bytes;
   final MediaDimensions? dimensions;
   final String? blurHash;
+  final String originalFilename;
 }
 
 class MediaDimensions {
@@ -188,6 +190,7 @@ Future<List<SelectedFile>?> selectMedia({
         filePath: media.path,
         bytes: mediaBytes,
         dimensions: await dimensions,
+        originalFilename: media.name,
       );
     }));
   }
@@ -221,6 +224,7 @@ Future<List<SelectedFile>?> selectMedia({
       filePath: pickedMedia.path,
       bytes: mediaBytes,
       dimensions: await dimensions,
+      originalFilename: pickedMedia.name,
     ),
   ];
 }
@@ -271,6 +275,7 @@ Future<List<SelectedFile>?> selectFiles({
         storagePath: storagePath,
         filePath: isWeb ? null : file.path,
         bytes: file.bytes!,
+        originalFilename: file.name,
       );
     }));
   }
@@ -284,6 +289,7 @@ Future<List<SelectedFile>?> selectFiles({
       storagePath: storagePath,
       filePath: isWeb ? null : file.path,
       bytes: file.bytes!,
+      originalFilename: file.name,
     )
   ];
 }
@@ -304,7 +310,8 @@ List<SelectedFile> selectedFilesFromUploadedFiles(
               false,
               isMultiData ? index : null,
             ),
-            bytes: file.bytes!);
+            bytes: file.bytes!,
+            originalFilename: file.originalFilename);
       },
     ).toList();
 
