@@ -8,6 +8,7 @@ class FFUploadedFile {
     this.height,
     this.width,
     this.blurHash,
+    this.originalFilename = '',
   });
 
   final String? name;
@@ -15,10 +16,11 @@ class FFUploadedFile {
   final double? height;
   final double? width;
   final String? blurHash;
+  final String originalFilename;
 
   @override
   String toString() =>
-      'FFUploadedFile(name: $name, bytes: ${bytes?.length ?? 0}, height: $height, width: $width, blurHash: $blurHash,)';
+      'FFUploadedFile(name: $name, bytes: ${bytes?.length ?? 0}, height: $height, width: $width, blurHash: $blurHash, originalFilename: $originalFilename,)';
 
   String serialize() => jsonEncode(
         {
@@ -27,6 +29,7 @@ class FFUploadedFile {
           'height': height,
           'width': width,
           'blurHash': blurHash,
+          'originalFilename': originalFilename,
         },
       );
 
@@ -38,6 +41,7 @@ class FFUploadedFile {
       'height': serializedData['height'],
       'width': serializedData['width'],
       'blurHash': serializedData['blurHash'],
+      'originalFilename': serializedData['originalFilename'] ?? '',
     };
     return FFUploadedFile(
       name: data['name'] as String,
@@ -45,6 +49,7 @@ class FFUploadedFile {
       height: data['height'] as double?,
       width: data['width'] as double?,
       blurHash: data['blurHash'] as String?,
+      originalFilename: data['originalFilename'] as String,
     );
   }
 
@@ -55,6 +60,7 @@ class FFUploadedFile {
         height,
         width,
         blurHash,
+        originalFilename,
       );
 
   @override
@@ -64,5 +70,6 @@ class FFUploadedFile {
       bytes == other.bytes &&
       height == other.height &&
       width == other.width &&
-      blurHash == other.blurHash;
+      blurHash == other.blurHash &&
+      originalFilename == other.originalFilename;
 }
