@@ -1,5 +1,4 @@
 import '/backend/backend.dart';
-import '/backend/schema/structs/index.dart';
 import '/flutter_flow/flutter_flow_util.dart';
 import '/index.dart';
 import 'timetable_page_widget.dart' show TimetablePageWidget;
@@ -16,6 +15,11 @@ class TimetablePageModel extends FlutterFlowModel<TimetablePageWidget> {
   void updateTableToFixStruct(Function(WeeklyTimeTableStruct) updateFn) {
     updateFn(tableToFix ??= WeeklyTimeTableStruct());
   }
+
+  ///  State fields for stateful widgets in this page.
+
+  // Stores action output result for [Firestore Query - Query a collection] action in Icon widget.
+  StudyDBbackupAggRecord? aa;
 
   @override
   void initState(BuildContext context) {}
